+++ beradio/config.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# (c) 2015 Andreas Motl, Elmyra UG <andreas.motl@elmyra.de>
+from beradio.util import ConfigStoreJson
+
+class BERadioConfig(ConfigStoreJson):
+    appname = 'beradio'