+++ beradio/publish.py
@@ -1,11 +1,10 @@
-#!/usr/bin/python
 # -*- coding: utf-8 -*-
 # (c) 2015 Richard Pobering <einsiedlerkrebs@netfrag.org>
 # (c) 2015 Andreas Motl, Elmyra UG <andreas.motl@elmyra.de>
 import sys
 import random
-from beradio.protocol import get_protocol_class
 from mqtt import BERadioMQTTAdapter
+from beradio.network import protocol_factory
 
 """
 Read data in Bencode format from command line, decode and publish via MQTT.
@@ -25,7 +24,7 @@ class DataToMQTT(object):
     def __init__(self, mqtt_broker, mqtt_topic='hiveeyes', protocol=2):
         self.mqtt_broker = mqtt_broker
         self.mqtt_topic = mqtt_topic
-        self.protocol_class = get_protocol_class(protocol)
+        self.protocol_class = protocol_factory(protocol)
 
     def setup(self):
         try: